Add SavedServiceBasedSetupStepper
