+++ js/src/pages/onboarding/setup-stepper/saved-service-based-setup-stepper/optimize-campaign.js
@@ -1,0 +1,18 @@
+/**
+ * Internal dependencies
+ */
+import useTargetAudienceFinalCountryCodes from '~/hooks/useTargetAudienceFinalCountryCodes';
+import CampaignAssetsForm from '~/components/paid-ads/campaign-assets-form';
+import AssetGroup from '~/components/paid-ads/asset-group';
+
+const OptimizeCampaign = ( { onSubmit } ) => {
+	const { data: countryCodes } = useTargetAudienceFinalCountryCodes();
+
+	return (
+		<CampaignAssetsForm onSubmit={ onSubmit } countryCodes={ countryCodes }>
+			<AssetGroup />
+		</CampaignAssetsForm>
+	);
+};
+
+export default OptimizeCampaign;
